requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.6.0
anthropic==0.34.2
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appnope==0.1.3
asgiref==3.8.1
astroid==3.3.4
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.2
build==1.2.2
cachetools==5.5.0
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
chroma-hnswlib==0.7.6
chromadb==0.5.11
click==8.1.7
coloredlogs==15.0.1
comm==0.1.4
contourpy==1.2.1
cryptography==43.0.1
cycler==0.12.1
dataclasses-json==0.6.7
datasets==2.14.5
debugpy==1.8.0
decorator==5.1.1
deepdiff==8.0.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.7
distro==1.9.0
docopt==0.6.2
effdet==0.4.1
einops==0.8.0
emoji==2.13.2
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.115.0
ffmpeg==1.4
ffmpy==0.4.0
filelock==3.12.4
filetype==1.2.0
flatbuffers==24.3.25
flit_core==3.9.0
fonttools==4.53.0
frozenlist==1.4.0
fsspec==2023.6.0
future==0.18.3
google-auth==2.34.0
googleapis-common-protos==1.65.0
gpt4all==2.8.2
gradio==4.44.1
gradio_client==1.3.0
greenlet==3.0.0
grpcio==1.66.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.25.1
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib_resources==6.4.0
iopath==0.1.10
ipykernel==6.25.2
ipython==8.15.0
ipython-ngql==0.7.5
isort==5.13.2
jedi==0.19.0
Jinja2==3.1.2
jiter==0.5.0
joblib==1.3.2
Js2Py==0.74
jsonlines==4.0.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpickle==3.0.2
jsonpointer==2.4
jupyter_client==8.3.1
jupyter_core==5.3.1
kiwisolver==1.4.5
kubernetes==30.1.0
langchain==0.3.1
langchain-anthropic==0.2.1
langchain-chroma==0.1.4
langchain-community==0.3.1
langchain-core==0.3.7
langchain-experimental==0.3.2
langchain-huggingface==0.1.0
langchain-nomic==0.1.3
langchain-ollama==0.2.0
langchain-openai==0.2.0
langchain-text-splitters==0.3.0
langchain-unstructured==0.1.5
langchainhub==0.1.21
langdetect==1.0.9
langgraph==0.2.30
langgraph-checkpoint==1.0.14
langsmith==0.1.129
layoutparser==0.3.4
lazy_loader==0.3
libmagic==1.0
librosa==0.10.1
llama-index==0.8.47
llamaapi==0.1.36
llvmlite==0.43.0
loguru==0.7.2
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-enum==1.5.1
matplotlib==3.9.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mmh3==5.0.0
monotonic==1.6
more-itertools==10.1.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nebula3-python==3.4.0
neo4j==5.13.0
nest-asyncio==1.6.0
networkx==3.1
nltk==3.9.1
nomic==3.1.2
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
ollama==0.3.3
omegaconf==2.3.0
onnx==1.16.1
onnxruntime==1.15.1
openai==1.46.0
openai-whisper @ git+https://github.com/openai/whisper.git@0a60fcaa9b86748389a656aa013c416030287d47
opencv-python==4.10.0.82
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orderly-set==5.2.2
orjson==3.10.7
overrides==7.7.0
packaging==24.1
pandas==2.1.1
parso==0.8.3
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pexpect==4.8.0
pi_heif==0.18.0
pickleshare==0.7.5
pikepdf==9.3.0
pillow==10.3.0
pillow_heif==0.18.0
pipwin==0.5.2
platformdirs==3.10.0
pooch==1.7.0
poppler-utils==0.1.0
portalocker==2.10.1
posthog==3.6.6
prompt-toolkit==3.0.39
protobuf==4.25.5
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==13.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.13
pycocotools==2.0.8
pycparser==2.21
pycryptodome==3.21.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.16.1
pyjsparser==2.7.1
PyJWT==2.9.0
pylint==3.3.1
pyparsing==3.1.1
pypdf==5.0.1
pypdfium2==4.30.0
PyPika==0.48.9
PyPrind==2.11.3
pyproject_hooks==1.1.0
pySmartDL==1.3.4
pytesseract==0.3.13
python-dateutil==2.8.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-magic-bin==0.4.14
python-multipart==0.0.12
python-oxmsg==0.0.1
pytz==2023.3.post1
pyvis==0.3.2
PyYAML==6.0.1
pyzmq==25.1.1
RapidFuzz==3.10.0
regex==2023.8.8
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.8.1
rsa==4.9
ruff==0.6.8
safetensors==0.4.5
scikit-learn==1.3.1
scipy==1.11.2
seaborn==0.13.2
semantic-version==2.10.0
sentence-transformers==3.1.1
sentencepiece==0.1.99
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.6
SpeechRecognition==3.10.4
SQLAlchemy==2.0.22
stack-data==0.6.2
starlette==0.38.5
sympy==1.12
tabulate==0.9.0
tavily-python==0.5.0
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.7.0
timm==1.0.9
tokenizers==0.20.0
tomli==2.0.1
tomlkit==0.12.0
torch==2.2.2
torchaudio==2.0.2
torchvision==0.17.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
transformers==4.45.1
typer==0.12.5
types-requests==2.32.0.20240914
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.1
ultralytics==8.2.32
ultralytics-thop==0.2.8
unstructured==0.11.5
unstructured-client==0.25.9
unstructured-inference==0.7.19
unstructured.pytesseract==0.3.13
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
wcwidth==0.2.6
websocket-client==1.8.0
websockets==12.0
whisper==1.1.10
wikipedia==1.4.0
wrapt==1.16.0
xxhash==3.3.0
yarl==1.9.2
zipp==3.17.0
